update averageComputing from LongWritable to DoubleWritable

+++ src/main/java/averageComputing/AverageComputingMapper.java
@@ -2,7 +2,7 @@ package averageComputing;
 
 import accessLog.Parser;
 import accessLog.Record;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -10,13 +10,13 @@ import org.apache.hadoop.mapreduce.Mapper;
 import java.io.IOException;
 
 public class AverageComputingMapper extends
-        Mapper<LongWritable, Text, Text, LongWritable> {
+        Mapper<LongWritable, Text, Text, DoubleWritable> {
     @Override
     protected void map(LongWritable key, Text record, Context context) throws IOException, InterruptedException {
         Record r = Parser.parserRecord(record);
         if(r != null) {
             Text text = new Text(r.getIpAddress());
-            LongWritable value = new LongWritable(r.getQuantity());
+            DoubleWritable value = new DoubleWritable(r.getQuantity());
             context.write(text, value);
         }
     }
