+++ src/main/java/averageComputing/AverageComputingReducer.java
@@ -1,6 +1,6 @@
 package averageComputing;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -8,16 +8,16 @@ import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
 
 public class AverageComputingReducer extends
-        Reducer<Text, LongWritable, Text, LongWritable> {
+        Reducer<Text, DoubleWritable, Text, DoubleWritable> {
     @Override
-    protected void reduce(Text word, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
-        Long sum = 0L;
+    protected void reduce(Text word, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
+        Double sum = 0D;
         int count = 0;
-        for (LongWritable value : values) {
+        for (DoubleWritable value : values) {
             sum += value.get();
             count += 1;
         }
-        Long avg = sum/count;
-        context.write(word, new LongWritable(avg));
+        Double avg = sum/count;
+        context.write(word, new DoubleWritable(avg));
     }
 }
